fix : read/write, group/group_member 분리
