--- backend/core/src/main/java/site/timecapsulearchive/core/domain/group/repository/memberGroupRepository/MemberGroupQueryRepository.java
@@ -1,10 +1,0 @@
-package site.timecapsulearchive.core.domain.group.repository.memberGroupRepository;
-
-import java.util.Optional;
-import site.timecapsulearchive.core.domain.group.data.dto.GroupOwnerSummaryDto;
-
-public interface MemberGroupQueryRepository {
-
-    Optional<GroupOwnerSummaryDto> findOwnerInMemberGroup(Long groupId, Long memberId);
-
-}
