--- backend/core/src/main/java/site/timecapsulearchive/core/domain/group/service/write/GroupWriteService.java
@@ -1,16 +1,0 @@
-package site.timecapsulearchive.core.domain.group.service.write;
-
-import site.timecapsulearchive.core.domain.group.data.dto.GroupCreateDto;
-
-public interface GroupWriteService {
-
-    void createGroup(final Long memberId, final GroupCreateDto dto);
-
-    void inviteGroup(final Long memberId, final Long groupId, final Long targetId);
-
-    void rejectRequestGroup(final Long memberId, final Long groupId, final Long targetId);
-
-    void acceptGroupInvite(final Long memberId, final Long groupId, final Long targetId);
-
-    void deleteGroup(final Long memberId, final Long groupId);
-}
