+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/group_member/data/GroupOwnerSummaryDto.java
@@ -1,4 +1,4 @@
-package site.timecapsulearchive.core.domain.group.data.dto;
+package site.timecapsulearchive.core.domain.group_member.data;
 
 public record GroupOwnerSummaryDto(
     String nickname,
