+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/group_member/entity/GroupInvite.java
@@ -1,4 +1,4 @@
-package site.timecapsulearchive.core.domain.group.entity;
+package site.timecapsulearchive.core.domain.group_member.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -12,6 +12,7 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import site.timecapsulearchive.core.domain.group.entity.Group;
 import site.timecapsulearchive.core.domain.member.entity.Member;
 import site.timecapsulearchive.core.global.entity.BaseEntity;
