+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/group_member/repository/groupInviteRepository/GroupInviteRepository.java
@@ -1,11 +1,11 @@
-package site.timecapsulearchive.core.domain.group.repository.groupInviteRepository;
+package site.timecapsulearchive.core.domain.group_member.repository.groupInviteRepository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.Repository;
 import org.springframework.data.repository.query.Param;
-import site.timecapsulearchive.core.domain.group.entity.GroupInvite;
+import site.timecapsulearchive.core.domain.group_member.entity.GroupInvite;
 
 public interface GroupInviteRepository extends Repository<GroupInvite, Long>,
     GroupInviteQueryRepository {
