+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/group_member/repository/memberGroupRepository/MemberGroupQueryRepository.java
@@ -1,0 +1,10 @@
+package site.timecapsulearchive.core.domain.group_member.repository.memberGroupRepository;
+
+import java.util.Optional;
+import site.timecapsulearchive.core.domain.group_member.data.GroupOwnerSummaryDto;
+
+public interface MemberGroupQueryRepository {
+
+    Optional<GroupOwnerSummaryDto> findOwnerInMemberGroup(Long groupId, Long memberId);
+
+}
