+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/group_member/repository/memberGroupRepository/MemberGroupRepository.java
@@ -1,8 +1,8 @@
-package site.timecapsulearchive.core.domain.group.repository.memberGroupRepository;
+package site.timecapsulearchive.core.domain.group_member.repository.memberGroupRepository;
 
 import java.util.List;
 import org.springframework.data.repository.Repository;
-import site.timecapsulearchive.core.domain.group.entity.MemberGroup;
+import site.timecapsulearchive.core.domain.group_member.entity.MemberGroup;
 
 public interface MemberGroupRepository extends Repository<MemberGroup, Long>,
     MemberGroupQueryRepository {
